+++ packages/core/src/commands/index.ts
@@ -5,6 +5,7 @@ export * from './insert-content';
 export * from './remove-content';
 export * from './append-blocks';
 export * from './change-property';
+export * from './change-style';
 export * from './change-title';
 export * from './change-value';
 export * from './move';