+++ packages/extension-add-block-menu/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chamaeleon/extension-add-block-menu
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @chamaeleon/core@0.5.0
+
 ## 0.0.8
 
 ### Patch Changes