+++ packages/extension-persist/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chamaeleon/extension-persist
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @chamaeleon/core@0.5.0
+
 ## 0.1.5
 
 ### Patch Changes